+++ cmake/PyneMacros.cmake
@@ -146,14 +146,20 @@ endmacro()
 
 # determine if spatial solver module should be built
 macro(pyne_set_build_spatial_solver)
+  SET(BUILD_SPATIAL_SOLVER false)
+  MESSAGE("-- Checking whether to build spatial solvers")
+  MESSAGE("-- -- Checking CMAKE_CXX_COMPILER_ID: ${CMAKE_CXX_COMPILER_ID}")
   IF(CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
+    MESSAGE("-- -- -- Checking CMAKE_CXX_COMPILER_VERSION: ${CMAKE_CXX_COMPILER_VERSION}")
+    MESSAGE("-- -- -- Checking if APPLE: ${APPLE}}")
     IF(NOT CMAKE_CXX_COMPILER_VERSION VERSION_LESS "4.6" AND
-       NOT APPLE AND NOT "$ENV{CONDA_BUILD}")
+       NOT APPLE )
       SET(BUILD_SPATIAL_SOLVER true)
     ELSE()
       SET(BUILD_SPATIAL_SOLVER false)
     ENDIF()
   ENDIF()
+  MESSAGE("-- Build spatial solvers: ${BUILD_SPATIAL_SOLVER}")
 endmacro()